main_thought_stream/requirement.txt
Django==5.1.4
psycopg2==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
